+++ Assets/lilToon/CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Z Bias` to outline
 - `Shadow Caster Bias` to rendering settings
 - Fur mesh division type
-- `GSAA` to reflection
+- `GSAA` and `Blending Mode` to reflection
+- `LOD` to strength / blur / AO mask of shadows
 
 ### Changed
 - Improved alpha mask GUI